Refactor
Extract Scriptableobject and SubModuleInfo Fetcher
