+++ Assets/Editor/SubModuleInfoFetcher.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9b73aa564e081d6458b84abaa7fd128d
